% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{fromGenoToMLdata}
\alias{fromGenoToMLdata}
\title{Function that obtains, from the genotype data, a dataset with the data
prepared for the Machine Learning part, after doing the variable selection}
\usage{
fromGenoToMLdata(workPath,
  path2Geno = "/home/users/gsit/juanbot/JUAN_SpanishGWAS/UNRELATED.SPAIN4.HARDCALLS.Rsq0.8",
  path2Covs = "/home/users/gsit/juanbot/JUAN_SpanishGWAS/COVS_SPAIN",
  predictor = "DISEASE",
  path2GWAS = "/home/users/gsit/juanbot/JUAN_SpanishGWAS/toJuanNov7th2018/",
  path2PRSice = "/home/users/gsit/juanbot/genoml-core/otherPackages/",
  path2plink = "", path2Pheno = paste0(workPath, "/MyPhenotype"),
  snpsSpain = NULL, iter)
}
\arguments{
\item{workPath}{String with your work path}

\item{path2Geno}{Path to your genotype data}

\item{path2Covs}{Path to your covariates data}

\item{predictor}{What you want to predict}

\item{path2GWAS}{Path to your GWAS}

\item{path2plink}{Path to PLINK}

\item{iter}{}

\item{path2PRSICE}{Path to PRSICE}

\item{path2pheno}{Path to phenotype data (optional)}
}
\value{
A handler with the mldata
}
\description{
Function that obtains, from the genotype data, a dataset with the data
prepared for the Machine Learning part, after doing the variable selection
}
\examples{
handlerMLdata <- fromGenoToMLdata("/home/rafael",
                                  "/home/users/gsit/juanbot/JUAN_SpanishGWAS/UNRELATED.SPAIN4.HARDCALLS.Rsq0.8",
                                  "/home/users/gsit/juanbot/JUAN_SpanishGWAS/COVS_SPAIN",
                                  "DISEASE",
                                  "/home/users/gsit/juanbot/JUAN_SpanishGWAS/toJuanNov7th2018/",
                                  "/home/users/gsit/juanbot/genoml-core/otherPackages/")
}
