% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_dev.R
\name{genDataFileNames}
\alias{genDataFileNames}
\title{Title Creating your genotipe data}
\usage{
genDataFileNames(genoHandler, onlyFold = -1,
  path2plink = "~/genoml-core-master/otherPackages/",
  which.to.create = c("train", "test"))
}
\arguments{
\item{which.to.create}{}
}
\description{
Title Creating your genotipe data
}
