% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_dev.R
\name{genDataFromHandler}
\alias{genDataFromHandler}
\title{Title Creating your genotipe data}
\usage{
genDataFromHandler(genoHandler, onlyFold = -1,
  which.to.create = c("train", "test"), lazy = F)
}
\arguments{
\item{which.to.create}{}
}
\description{
Title Creating your genotipe data
}
