% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{genModels}
\alias{genModels}
\title{Title}
\usage{
genModels(algs, handlerMLdata, imputeMissingData, workPath, gridSearch)
}
\arguments{
\item{algs}{Algorithms we want to use for the obtaining of the models}

\item{handlerMLdata}{handler that contains the *.dataForML files with the data}

\item{imputeMissingData}{string with the type of preprocess we want to apply to the data}

\item{workPath}{String with your work path}

\item{gridSearch}{int with the gridSearch for the caret train method}
}
\value{
the models produced
}
\description{
Title
}
\examples{
models = genModels(c("glm","nnet","rf","xgbTree"), handler, "median", "/home/rafael", 30)
}
