% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_dev.R
\name{genoMLtrainAndTest}
\alias{genoMLtrainAndTest}
\title{Title}
\usage{
genoMLtrainAndTest(handler, outcome = c("cont", "disc"),
  skipMLinit = F, ncores = 1, learningAlgorithm = "xgbTree",
  cvReps = 10, gridSearch = 30, imputeMissingData = "median",
  useCluster = T, clLogPath = "~/launch/",
  caretPath = "/home/jbotia/caret/pkg/caret/",
  clParams = " -l h_rt=96:0:0 -l tmem=3G,h_vmem=3G ")
}
\arguments{
\item{clParams}{}
}
\description{
Title
}
