% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_dev.R
\name{getHandlerToGenotypeData}
\alias{getHandlerToGenotypeData}
\title{Title Getting a wrapper to plink files without using them directly
A handler is a wrapper to plink files and it is useful to isolate
the developer of managing such files so he/she can be focused on
designing the ML pipeline with the right data. Besides, having a handler
to the data does not necessarily mean you have the plink files ready.
It only mean that if not, the handler knows how to create them. In this
way you optimize disk.}
\usage{
getHandlerToGenotypeData(geno, covs, predictor, id, fid, pheno,
  verify = T, wp = dirname(pheno))
}
\arguments{
\item{geno}{Full path to the plink files, not including the extension.
For example, for /home/myuser/mygeno.(bed/bim/fam) files, the geno value
should be the common prefix "/home/myuser/mygeno".}

\item{covs}{A covariates file with the minimal information about samples}

\item{pheno}{}
}
\description{
Title Getting a wrapper to plink files without using them directly
A handler is a wrapper to plink files and it is useful to isolate
the developer of managing such files so he/she can be focused on
designing the ML pipeline with the right data. Besides, having a handler
to the data does not necessarily mean you have the plink files ready.
It only mean that if not, the handler knows how to create them. In this
way you optimize disk.
}
