% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_dev.R
\name{getPartitionsFromHandler}
\alias{getPartitionsFromHandler}
\title{Title}
\usage{
getPartitionsFromHandler(genoHandler, path2plink, workPath = NULL,
  how = "k-fold cv", k = 10, p = 0.75)
}
\arguments{
\item{genoHandler}{A handler obtained by calling to `getHandlerToGenotypeData()`}

\item{how}{We can get a sample with "holdout" and "k-fold cv"}

\item{k}{The number of folds when we use cross-validation}

\item{p}{The proportion of data that does to training}
}
\value{
The same handler, with two new fields, `trainFolds` and `testFolds` with the indexes
}
\description{
Title
}
