% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_dev.R
\name{getPartitionsFromMLData}
\alias{getPartitionsFromMLData}
\title{Title}
\usage{
getPartitionsFromMLData(mldatafile, workPath = NULL, how = "k-fold cv",
  k = 10, p = 0.75)
}
\arguments{
\item{how}{We can get a sample with "holdout" and "k-fold cv"}

\item{k}{The number of folds when we use cross-validation}

\item{p}{The proportion of data that does to training}

\item{genoHandler}{A handler obtained by calling to `getHandlerToGenotypeData()`}
}
\value{
The same handler, with two new fields, `trainFolds` and `testFolds` with the indexes
}
\description{
Title
}
