% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{metaLaunch}
\alias{metaLaunch}
\title{Run the complete meta-learning pipeline}
\usage{
metaLaunch(workPath, trainSpeed, includeGeno, imputeMissingData,
  gridSearch)
}
\arguments{
\item{workPath}{String with your work path}

\item{trainSpeed}{type of models we want to run in our metalearning function, can be FAST, FURIOUS, ALL OR BOOSTED}

\item{includeGeno}{bool indicating if we are going to include genotype in the meta-learning dataset}

\item{imputeMissingData}{string with the type of preprocess we want to apply to the data}

\item{gridSearch}{int with the gridSearch for the caret train method}

\item{handlerML}{string with the location of the handler that contains the *.dataForML files with the data}
}
\value{
A handler with the meta learning results
}
\description{
First we obtain the models calling to genModels and then
we pass those models to the metaLearning function, which will do the rest
}
\examples{
handlerMeta = metaLaunch("/home/rafael", "FAST", F, "median", 30)
}
