% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{metaMLtrainAndTest}
\alias{metaMLtrainAndTest}
\title{Meta - Learning function}
\usage{
metaMLtrainAndTest(repoModels, handlersML, alg = "xgbTree",
  workPath = "/home/rafa/MetaLearning/metaML/Modelos/",
  includeGeno = F, gridSearch = 30, imputeMissingData = "median")
}
\arguments{
\item{repoModels}{List containing best models for each repository}

\item{handlersML}{List containing the ML-handlers associated to each one of the repositories}

\item{alg}{the algorithm we are going to use to train the meta-dataset}

\item{workPath}{String with your work path}

\item{includeGeno}{bool indicating if we are going to include genotype in the meta-learning dataset}

\item{gridSearch}{int with the gridSearch for the caret train method}

\item{imputeMissingData}{string with the type of preprocess we want to apply to the data}
}
\value{
a handler with all the meta-learning results
}
\description{
Creating the meta machine learning data and doing
the learning on this new dataset.
}
\examples{
handlerMeta = metaMLtrainAndTest(repoModels, handlersML, "xgbTree", "/home/rafael", F, 30, "median")
}
