% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_dev.R
\name{mostRelevantSNPs}
\alias{mostRelevantSNPs}
\title{Title}
\usage{
mostRelevantSNPs(handler, reduce = "PRSICE", phenoScale = "DISC",
  gwas = "RISK_noSpain.tab", SNPcolumnatGWAS = "SNP", herit = NA,
  clumpField = "p", path2plink, addit = "NA",
  path2gcta64 = path2plink, cores = 1, prune_windowSize = 10000,
  prune_stepSize = 1, prune_r2 = 0.1, path2PRSice = path2plink,
  PRSiceexe = "PRSice_linux", workPath = NULL, path2GWAS, force = F)
}
\arguments{
\item{path2plink}{}
}
\description{
Title
}
