% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_dev.R
\name{verifyHandler}
\alias{verifyHandler}
\title{Title Verifies whether the handler has data created}
\usage{
verifyHandler(h, level = 1)
}
\arguments{
\item{level}{}
}
\description{
Title Verifies whether the handler has data created
}
