% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{metaLaunch}
\alias{metaLaunch}
\title{Run the complete meta-learning pipeline}
\usage{
metaLaunch(workPath, includeGeno, imputeMissingData, gridSearch, nrepos)
}
\arguments{
\item{workPath}{String with your work path}

\item{includeGeno}{bool indicating if we are going to include genotype in the meta-learning dataset}

\item{imputeMissingData}{string with the type of preprocess we want to apply to the data}

\item{gridSearch}{int with the gridSearch for the caret train method}

\item{nrepos}{Number of repositories we are working with}
}
\value{
A handler with the meta learning results
}
\description{
In a first step, we generate two datasets, for train and test,
and then we obtain the best model applying ML to the train dataset, for each repository.
}
\details{
Once we have the best models, we run the meta learning on them to generate a more reliable result
}
\examples{
handlerMeta = metaLaunch("/home/rafael", F, "median", 30, 3)
}
