% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{metaMLtrainAndTest}
\alias{metaMLtrainAndTest}
\title{Meta - Learning function}
\usage{
metaMLtrainAndTest(models, handlerMLdata, algs, trainSpeed = "ALL",
  workPath = "/home/rafa/MetaLearning/metaML/Modelos/",
  includeGeno = F, gridSearch = 30, imputeMissingData = "median")
}
\arguments{
\item{models}{Models obtained in the function genModels}

\item{handlerMLdata}{handler that contains the *.dataForML files with the data}

\item{algs}{the algorithms we are going to use to train the meta-dataset}

\item{trainSpeed}{type of models we want to run in our metalearning function}

\item{workPath}{String with your work path}

\item{includeGeno}{bool indicating if we are going to include genotype in the meta-learning dataset}

\item{gridSearch}{int with the gridSearch for the caret train method}

\item{imputeMissingData}{string with the type of preprocess we want to apply to the data}
}
\value{
a handler with all the meta-learning results
}
\description{
Creating the meta machine learning data and doing
the learning on this new dataset.
}
\examples{
handlerMeta = metaMLtrainAndTest(models, handler, c("glm","nnet","rf","xgbTree"), "/home/rafael", F, 30, "median")
}
